Corrected ssh option handling
The option ip_v6_disabled and the undocumented option exec was incorrectly
handled by the option checking code.

There where two options for allowing user interaction one documented
and one undocumented, they where merge to to be the documented one.
For backwards compatibility the undocumented one is transformed to the
documented one in the option checking code.