SSH: Make "auth_methods" server option re-usable
The 'auth_methods' option is used by the server side of the SSH code to
tell a connecting SSH client about the authentication methods that are
supported by the server. The code still extracts and handles the
'auth_methods' option from Opts in appropriate places, but the Opts
checking code in ssh.erl didn't allow that option to be specified.